requirements.txt
sounddevice
numpy
openai-whisper
nltk
Flask
Flask-Cors
